requirements.txt
psycopg2
requests
kafka-python
apache-airflow
apache-airflow-providers-postgres